+++ api/Controllers/StockController.cs
@@ -1,4 +1,5 @@
 using api.Dtos.Stock;
+using api.Helpers;
 using api.Interfaces;
 using api.Mappers;
 using api.Models;
@@ -19,11 +20,14 @@ public class StockController : ControllerBase
         this.stockRepo = stockRepo;
     }
 
-    //* GET api/stock
+    //* GET api/stock?symbol=ABC&companyName=Company
     [HttpGet]
-    public async Task<IActionResult> GetAll()
+    public async Task<IActionResult> GetAll([FromQuery] StockQuery query)
     {
-        var stocks = await stockRepo.GetAllAsync();
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+
+        var stocks = await stockRepo.GetAllAsync(query);
 
         var stockDto = stocks.Select(s => s.ToStockDto());
 
@@ -31,15 +35,15 @@ public class StockController : ControllerBase
     }
 
     //* GET api/stock/{id}
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
         var stock = await stockRepo.GetByIdAsync(id);
 
         if (stock == null)
-        {
             return NotFound("Stock not found");
-        }
 
         return Ok(stock.ToStockDto());
     }
@@ -47,6 +51,9 @@ public class StockController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateStockDto stockDto)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+
         var stockModel = stockDto.ToStockFromCreateDTO();
 
         await stockRepo.CreateAsync(stockModel);
@@ -60,30 +67,32 @@ public class StockController : ControllerBase
 
     //* PUT api/stock/{id}
     [HttpPut]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] CreateStockDto updateDto)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+
         var stockModel = await stockRepo.UpdateAsync(id, updateDto);
 
         if (stockModel == null)
-        {
             return NotFound("Stock not found");
-        }
 
         return Ok(stockModel.ToStockDto());
     }
 
     //* DELETE api/stock/{id}
     [HttpDelete]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+
         var stockModel = await stockRepo.DeleteAsync(id);
 
         if (stockModel == null)
-        {
             return NotFound("Stock not found");
-        }
 
         return NoContent();
     }
